+++ .gitignore
@@ -384,3 +384,4 @@ _UpgradeReport_Files/
 Thumbs.db
 Desktop.ini
 .DS_Store
+.idea/
