+++ USFMConverter/UI/MainWindow.axaml
@@ -8,7 +8,7 @@
         Title="USFMConverter"
         Background="AliceBlue">
   <StackPanel>
-    <screen:HomeScreen Name="HomeScreen" FolderSelected="HomeScreen_FolderSelected"/>
-    <screen:FilesScreen Name="FilesScreen" IsVisible="False"/>
+    <!-- <screen:HomeScreen Name="HomeScreen" FolderSelected="HomeScreen_FolderSelected"/> -->
+    <screen:ProjectDetailScreen Name="ProjectDetailScreen" IsVisible="False"/>
   </StackPanel>
 </Window>
