--- USFMConverter/UI/Pages/FormatScreen.axaml.cs
@@ -1,19 +1,0 @@
-using Avalonia;
-using Avalonia.Controls;
-using Avalonia.Markup.Xaml;
-
-namespace USFMConverter.UI.Pages
-{
-    public partial class FormatScreen : UserControl
-    {
-        public FormatScreen()
-        {
-            InitializeComponent();
-        }
-
-        private void InitializeComponent()
-        {
-            AvaloniaXamlLoader.Load(this);
-        }
-    }
-}
